Move Abbreviations to a separate module
