+++ src/cutty/core/config.py
@@ -4,43 +4,14 @@ from __future__ import annotations
 from dataclasses import dataclass
 from dataclasses import field
 from pathlib import Path
-from typing import Mapping
 from typing import Optional
 
 import poyo.exceptions
 
 from . import exceptions
 from . import locations
+from .abbreviations import Abbreviations
 from .utils import with_context
-
-
-class Abbreviations:
-    """Abbreviations for template locations."""
-
-    def __init__(self) -> None:
-        """Initialize."""
-        self.abbreviations = {
-            "gh": "https://github.com/{}.git",
-            "gl": "https://gitlab.com/{}.git",
-            "bb": "https://bitbucket.org/{}",
-        }
-
-    def update(self, abbreviations: Mapping[str, str]) -> None:
-        """Update the registered abbreviations."""
-        self.abbreviations.update(abbreviations)
-
-    def expand(self, template: str) -> str:
-        """Expand abbreviations in a template location."""
-        if template in self.abbreviations:
-            return self.abbreviations[template]
-
-        prefix, _, rest = template.partition(":")
-        abbreviation = self.abbreviations.get(prefix)
-        if abbreviation is not None:
-            with exceptions.InvalidAbbreviation(prefix, template, abbreviation):
-                return abbreviation.format(rest)
-
-        return template
 
 
 @dataclass
